separate  mod fileitem and mod diritem
